remove simple-build from build tests
